refactor: integrate stress tests into module test files
- Moved all stress tests from stress.test.ts into their respective
  module test files (quotes.test.ts, dashes.test.ts, symbols.test.ts,
  index.test.ts)
- Converted "known limitations" tests to proper it.failing() tests
  that document bugs:
  - Empty single quotes ('') → should be left+right pair
  - Whitespace-only quotes (" ") → should convert both quotes
  - Chained multiplications (5x5x5) → should convert all x's
  - Short phone numbers (555-1234) → should preserve hyphens
- Updated README with newly discovered known limitations
- Deleted stress.test.ts after integration
- Total: 762 tests, 100% coverage
