+++ src/tests/quotes.test.ts
@@ -219,6 +219,76 @@ describe("niceQuotes", () => {
     })
   })
 
+  describe("unbalanced and edge quotes", () => {
+    it.each([
+      ['"unclosed', `${LEFT_DOUBLE_QUOTE}unclosed`],
+      ["unclosed'", `unclosed${RIGHT_SINGLE_QUOTE}`],
+      ['""', `${LEFT_DOUBLE_QUOTE}${RIGHT_DOUBLE_QUOTE}`],
+    ])('handles edge quote pattern: "%s"', (input, expected) => {
+      expect(niceQuotes(input)).toBe(expected)
+    })
+  })
+
+  describe("quotes after various punctuation", () => {
+    it.each([
+      ['text; "quote"', `text; ${LEFT_DOUBLE_QUOTE}quote${RIGHT_DOUBLE_QUOTE}`],
+      ['title: "quote"', `title: ${LEFT_DOUBLE_QUOTE}quote${RIGHT_DOUBLE_QUOTE}`],
+      ['what? "quote"', `what? ${LEFT_DOUBLE_QUOTE}quote${RIGHT_DOUBLE_QUOTE}`],
+      ['wow! "quote"', `wow! ${LEFT_DOUBLE_QUOTE}quote${RIGHT_DOUBLE_QUOTE}`],
+      ['[note] "quote"', `[note] ${LEFT_DOUBLE_QUOTE}quote${RIGHT_DOUBLE_QUOTE}`],
+      ['{code} "quote"', `{code} ${LEFT_DOUBLE_QUOTE}quote${RIGHT_DOUBLE_QUOTE}`],
+    ])('handles quotes after punctuation: "%s"', (input, expected) => {
+      expect(niceQuotes(input)).toBe(expected)
+    })
+  })
+
+  describe("consecutive contractions", () => {
+    it.each([
+      ["I'd've", `I${RIGHT_SINGLE_QUOTE}d${RIGHT_SINGLE_QUOTE}ve`],
+      ["y'all'd've", `y${RIGHT_SINGLE_QUOTE}all${RIGHT_SINGLE_QUOTE}d${RIGHT_SINGLE_QUOTE}ve`],
+      ["wouldn't've", `wouldn${RIGHT_SINGLE_QUOTE}t${RIGHT_SINGLE_QUOTE}ve`],
+      ["couldn't've", `couldn${RIGHT_SINGLE_QUOTE}t${RIGHT_SINGLE_QUOTE}ve`],
+    ])('handles double contraction: "%s"', (input, expected) => {
+      expect(niceQuotes(input)).toBe(expected)
+    })
+  })
+
+  describe("quotes with special Unicode", () => {
+    it.each([
+      ['"Hello! 😊"', `${LEFT_DOUBLE_QUOTE}Hello! 😊${RIGHT_DOUBLE_QUOTE}`],
+      ['"你好"', `${LEFT_DOUBLE_QUOTE}你好${RIGHT_DOUBLE_QUOTE}`],
+      ['"こんにちは"', `${LEFT_DOUBLE_QUOTE}こんにちは${RIGHT_DOUBLE_QUOTE}`],
+      ['"مرحبا"', `${LEFT_DOUBLE_QUOTE}مرحبا${RIGHT_DOUBLE_QUOTE}`],
+      ['"Привет"', `${LEFT_DOUBLE_QUOTE}Привет${RIGHT_DOUBLE_QUOTE}`],
+    ])('handles Unicode content in quotes: "%s"', (input, expected) => {
+      expect(niceQuotes(input)).toBe(expected)
+    })
+  })
+
+  describe("multiline quotes", () => {
+    it.each([
+      ['"Hello\nWorld"', `${LEFT_DOUBLE_QUOTE}Hello\nWorld${RIGHT_DOUBLE_QUOTE}`],
+      ['"A"\n"B"', `${LEFT_DOUBLE_QUOTE}A${RIGHT_DOUBLE_QUOTE}\n${LEFT_DOUBLE_QUOTE}B${RIGHT_DOUBLE_QUOTE}`],
+    ])('handles multiline: "%s"', (input, expected) => {
+      expect(niceQuotes(input)).toBe(expected)
+    })
+  })
+
+  describe("split dialogue", () => {
+    it.each([
+      [
+        '"Yes," he said, "absolutely."',
+        `${LEFT_DOUBLE_QUOTE}Yes,${RIGHT_DOUBLE_QUOTE} he said, ${LEFT_DOUBLE_QUOTE}absolutely.${RIGHT_DOUBLE_QUOTE}`,
+      ],
+      [
+        '"No," she replied, "I disagree."',
+        `${LEFT_DOUBLE_QUOTE}No,${RIGHT_DOUBLE_QUOTE} she replied, ${LEFT_DOUBLE_QUOTE}I disagree.${RIGHT_DOUBLE_QUOTE}`,
+      ],
+    ])('handles split dialogue: "%s"', (input, expected) => {
+      expect(niceQuotes(input)).toBe(expected)
+    })
+  })
+
   describe("punctuationStyle option", () => {
     // Reusable test strings
     const periodOutsideDouble = `${LEFT_DOUBLE_QUOTE}Hello${RIGHT_DOUBLE_QUOTE}.`
@@ -268,6 +338,22 @@ describe("niceQuotes", () => {
       it("still converts straight quotes to smart quotes", () => {
         expect(niceQuotes('"Hello".', { punctuationStyle: "none" })).toBe(periodOutsideDouble)
       })
+    })
+  })
+
+  describe("known bugs", () => {
+    // BUG: Empty single quotes become double apostrophes instead of left+right single quotes
+    it.failing("should convert empty single quotes to left+right pair", () => {
+      expect(niceQuotes("''")).toBe(`${LEFT_SINGLE_QUOTE}${RIGHT_SINGLE_QUOTE}`)
+    })
+
+    // BUG: Quote containing only whitespace has limited conversion
+    it.failing("should convert quotes containing only whitespace", () => {
+      expect(niceQuotes('" "')).toBe(`${LEFT_DOUBLE_QUOTE} ${RIGHT_DOUBLE_QUOTE}`)
+    })
+
+    it.failing("should convert single quotes containing only whitespace", () => {
+      expect(niceQuotes("' '")).toBe(`${LEFT_SINGLE_QUOTE} ${RIGHT_SINGLE_QUOTE}`)
     })
   })
 })
